% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomads_filelist.R
\name{get_nomads_filelist}
\alias{get_nomads_filelist}
\title{Getting Most Current NWM data from the NOAA NOMADs server}
\usage{
get_nomads_filelist(dir = "./inst/flowlinefinder/data/current_nc",
  type = NULL, time = NULL, num = 6)
}
\value{
This fuction saves the 18 NetCDF files related to the most recent forecasts.
}
\description{
This function accesses the NOAA Nomads server to download the most up to date forecast NetCDF files.
These files cover the spatial domain of the continential United States. These files will be stored in the
'./FlowlineFinder' subfolder of the specified directory. Each time this function is run the old files will be removed.
So, if you see something interesting be sure to save the data elsewhere!
}
\author{
Mike Johnson
}
