% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_risk_map.R
\name{make_flood_risk_map}
\alias{make_flood_risk_map}
\title{Make Flood Risk Max}
\usage{
make_flood_risk_map(path = NULL, dir = NULL)
}
\arguments{
\item{path}{a path the the maximum value}
}
\value{
a leaflet map
}
\description{
Make Flood Risk Max
}
